LPS-90068 calendar-service: Fix package name https://issues.liferay.com/browse/LPS-90068
