+++ modules/apps/calendar/calendar-service/src/main/java/com/liferay/calendar/internal/search/spi/model/index/contributor/CalendarBookingModelIndexerWriterContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.calendar.internal.search;
+package com.liferay.calendar.internal.search.spi.model.index.contributor;
 
 import com.liferay.calendar.model.CalendarBooking;
 import com.liferay.calendar.service.CalendarBookingLocalService;
